Linker error with boost, trying to move file checkers to ioHandlers
